+++ Readme/CHANGELOG.md
@@ -10,6 +10,7 @@ If you would like to see fresh changes done since last release you can do so on 
 
 - [Changelog](#changelog)
   - [Table of Contents](#table-of-contents)
+  - [v0.11.1 (develop branch)](#v0111-develop-branch)
   - [v0.11.0 (current release)](#v0110-current-release)
   - [v0.10.0](#v0100)
   - [v0.9.0](#v090)
@@ -22,9 +23,15 @@ If you would like to see fresh changes done since last release you can do so on 
   - [v0.3.0](#v030)
   - [v0.2](#v02)
 
-## v0.11.0 (current release)
+## v0.11.1 (develop branch)
 
 **NOTE:** Changes for unreleased version (develop branch) may change or be announced upfront
+
+- Rules
+
+  - Added rules for ColorMania color picker
+
+## v0.11.0 (current release)
 
 - Rules